Develop

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import get_jwt_identity, jwt_required, get_jwt
 from models.item import ItemModel
 
 
@@ -16,13 +16,14 @@ class Item(Resource):
                         help="Every item needs a store_id."
                         )
 
-    @jwt_required()
-    def get(self, name: str):
-        item = ItemModel.find_by_name(name=name)
+    @jwt_required(optional=True)
+    def get(self, name):
+        item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @jwt_required(refresh=True)
     def post(self, name: str):
         if ItemModel.find_by_name(name=name):
             return {'message': "An item with name '{}' already exists.".format(name)}
@@ -38,13 +39,17 @@ class Item(Resource):
 
         return item.json()
 
+    @jwt_required(optional=True)
     def delete(self, name: str):
-        item = ItemModel.find_by_name(name=name)
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
 
+        item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {'message': 'Item deleted'}
+            return {'message': 'Item deleted.'}
+        return {'message': 'Item not found.'}, 404
 
     def put(self, name: str):
         data = Item.parser.parse_args()
@@ -62,7 +67,14 @@ class Item(Resource):
 
 class ItemList(Resource):
 
+    @jwt_required(optional=True)
     def get(self):
+        user_id = get_jwt_identity()
         # items = list(map(lambda item: item.json(), items))
         items = [item.json() for item in ItemModel.get_all()]
-        return {'items': items}
+        if user_id:
+            return {'items': items}, 200
+        return {
+                   'items': [item['name'] for item in items],
+                   'message': 'More data available if you log in.'
+               }, 200
